Finished attendance page
Graph completed and fixed small errors:
	Graph had int truncation changing values

Data View started and completed:
	Total hours over period of time graph is displaying
	Compliance score over period of time the graph is displaying

+++ src/util/DataPoint.java
@@ -16,9 +16,9 @@ import javax.swing.JPopupMenu;
 public class DataPoint extends JButton {
 
 	private String desc;
-	private int value;
+	private double value;
 
-	public DataPoint(String descriptor, int pointVal, String regex) {
+	public DataPoint(String descriptor, double pointVal, String regex) {
 
 		desc = descriptor;
 		value = pointVal;
@@ -53,7 +53,7 @@ public class DataPoint extends JButton {
 
 	}
 
-	public int getValue() {
+	public double getValue() {
 
 		return value;
